requirements.txt
# Core trading and data dependencies
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
ta-lib>=0.4.0

# Machine learning and AI
scikit-learn>=1.3.0
# tensorflow>=2.13.0  # Commented out for Python 3.13 compatibility
# torch>=2.0.0        # Commented out for Python 3.13 compatibility
# xgboost>=1.7.0      # Commented out for Python 3.13 compatibility

# Data visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=11.0

# Logging and monitoring
loguru>=0.7.0
prometheus-client>=0.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For advanced features
# lit-protocol>=0.1.0  # Uncomment when Lit Protocol integration is ready
# gaia-client>=0.1.0   # Uncomment when Gaia integration is ready
# mastra-client>=0.1.0 # Uncomment when Mastra integration is ready 